+++ src/main/java/com/kshrd/jpa_hibernate_homework/service/ProductService.java
@@ -16,7 +16,7 @@ public interface ProductService {
 
     List<Product> getProductByName(String name);
 
-    List<Product> getProductByQuantity(Long quantity);
+    List<Product> getProductByQuantity(Integer quantity);
 
     Product createProduct(@Valid ProductRequest request);
